# frontend/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sofix API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/players/{slug}/games": {
      "get": {
        "summary": "Player History",
        "description": "Local counterpart of the web server's two parameterized history queries.",
        "operationId": "player_history_api_players__slug__games_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__Any__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sorare/week/{slug}": {
      "get": {
        "summary": "Sorare Week",
        "description": "One finished gameweek the job kept whole, by its Sorare slug. Local development only, like /sorare.",
        "operationId": "sorare_week_api_sorare_week__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__Any__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sorare/alt/{slug}": {
      "get": {
        "summary": "Sorare Alt",
        "description": "The week planned again on Sorare's own projections, by its Sorare slug. Local development only, like /sorare.",
        "operationId": "sorare_alt_api_sorare_alt__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__Any__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sorare/ahead/{round}": {
      "get": {
        "summary": "Sorare Ahead",
        "description": "The early plan the job made for a LaLiga round Sorare has not opened. Local development only, like /sorare.",
        "operationId": "sorare_ahead_api_sorare_ahead__round__get",
        "parameters": [
          {
            "name": "round",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Round"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__Any__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_dict_str__Any__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GridMeta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ApiResponse[dict[str, Any]]"
      },
...
      "GridMeta": {
        "properties": {
          "last_synced_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Synced At"
          },
          "last_predicted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Predicted At"
          },
          "model_notes": {
            "items": {
              "$ref": "#/components/schemas/ModelNote"
            },
            "type": "array",
            "title": "Model Notes",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "last_synced_at",
          "last_predicted_at"
        ],
        "title": "GridMeta"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelNote": {
        "properties": {
          "lens": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "overall",
                  "attack",
                  "defence"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Lens"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "lens",
          "text"
        ],
        "title": "ModelNote"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}